+++ pennylane_lightning/lightning_qubit/_adjoint_jacobian.py
@@ -293,7 +293,16 @@ class LightningAdjointJacobian:
                 "Adjoint differentiation does not support State measurements."
             )
 
+        if any(m.return_type is not Expectation for m in tape.measurements):
+            raise QuantumFunctionError(
+                "Adjoint differentiation method does not support expectation return type "
+                "mixed with other return types"
+            )
+
         # Proceed, because tape_return_type is Expectation.
+        if qml.math.ndim(grad_vec) == 0:
+            grad_vec = (grad_vec,)
+
         if len(grad_vec) != len(measurements):
             raise ValueError(
                 "Number of observables in the tape must be the same as the "