+++ .github/CHANGELOG.md
@@ -24,6 +24,9 @@
   [(#607)](https://github.com/PennyLaneAI/pennylane-lightning/pull/607)
   [(#628)](https://github.com/PennyLaneAI/pennylane-lightning/pull/628)
 
+* Add Vector-Jacobian Product calculation support to `lightning.qubit`.
+  [(#644)](https://github.com/PennyLaneAI/pennylane-lightning/pull/644)
+
 * Add support for using new operator arithmetic as the default.
   [(#649)](https://github.com/PennyLaneAI/pennylane-lightning/pull/649)